Edit is_isomorphic.py, 205
