+++ data_structures_and_algorithms/hashing/addtl_probs/205_isomorphic_strings/input.json
@@ -13,5 +13,15 @@
         "s": "paper",
         "t": "title",
         "expected": true
+    },
+    {
+        "s": "badc",
+        "t": "baba",
+        "expected": false
+    },
+    {
+        "s": "papap",
+        "t": "titii",
+        "expected": false
     }
 ]
